Wrote a few tests around loading the 7z library, but they cause affect other tests, so I'm left with just one...
